♻️ Move some data helpers into datadog_common_test
The randomString and randomElement functions are likely to be useful across all packages for testing so better to move them into the common test utilities package.
