+++ packages/datadog_common_test/lib/src/data_helpers.dart
@@ -1,0 +1,24 @@
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2019-Present Datadog, Inc.
+import 'dart:math';
+
+final _random = Random();
+const _alphas = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+const _numerics = '0123456789';
+const _alphaNumerics = _alphas + _numerics;
+
+String randomString({int length = 10}) {
+  final result = String.fromCharCodes(Iterable.generate(
+    length,
+    (_) => _alphaNumerics.codeUnitAt(_random.nextInt(_alphaNumerics.length)),
+  ));
+
+  return result;
+}
+
+extension RandomExtension<T> on List<T> {
+  T randomElement() {
+    return this[_random.nextInt(length)];
+  }
+}
