+++ packages/datadog_flutter_plugin/e2e_test_app/integration_test/rum/rum_monitor_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:math';
 
+import 'package:datadog_common_test/datadog_common_test.dart';
 import 'package:datadog_flutter_plugin/datadog_flutter_plugin.dart';
 import 'package:e2e_test_app/main.dart' as app;
 import 'package:flutter_test/flutter_test.dart';
