+++ packages/datadog_flutter_plugin/e2e_test_app/pubspec.yaml
@@ -20,6 +20,8 @@ dev_dependencies:
   integration_test:
     sdk: flutter
   flutter_lints: ^1.0.0
+  datadog_common_test:
+    path: ../../datadog_common_test
 
 flutter:
   uses-material-design: true
