+++ packages/datadog_flutter_plugin/pubspec.yaml
@@ -22,6 +22,8 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^2.0.1
   mocktail: ^0.2.0
+  datadog_common_test:
+    path: ../datadog_common_test
 
 flutter:
   plugin:
